more comments on App.js
