+++ client/src/App.js
@@ -13,10 +13,21 @@ import "./styles/App.css";
 function App() {
 
 	const [selectedImg, setSelectedImg] = useState(null);
+	const [slides, prepSlideShow] = useState(null);
 
 	const setViewImg = (myImg) => {
-	//`	console.log("from setView: " + myImg + "\n");
 		setSelectedImg(myImg);
+		prepSlideShow(myImg)
+/*Now need to send slide to slideshowdata for setting up and return it as an array of images to FullScreenView.js
+		 **HINT**____________________________________________________________________________________
+		 	USE REDUCER.
+				This is because the slide show array (slides) needs to be accessed by
+				SlideShowData.js and SlideShowData is not a component. So we cannot wrap it with a
+				SlideShowData.Provider. SlideShowData is doesnt have return() so it's not a component
+
+			In reducer the array will be easy to set up. Use the same idea as basket set up
+			and take data in and out of the array.
+		 */ 
 	}
 
 	return(
